refactor: move VariablesEditor helpers to utils.js
